# app/Http/Controllers/QuotesController.php
<?php

namespace App\Http\Controllers;

use App\Models\Quotes;
use Illuminate\Http\Request;

class QuotesController extends Controller
{

     /**
     * @OA\Get (
     *     path="/api/quotes",
     *      operationId="all_quotes",
     *     tags={"quotes"},
     *     security={{ "apiAuth": {} }},
     *     summary="All quotes",
     *     description="All quotes",
     *     @OA\Parameter(
     *         in="path",
     *         name="id",
     *         required=true,
     *         @OA\Schema(type="string")
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="OK",
     *         @OA\JsonContent(
     *              @OA\Property(property="id", type="number", example=1),
     *              @OA\Property(property="name", type="string", example="Process"),
     *              @OA\Property(property="created_at", type="string", example="2023-02-23T00:09:16.000000Z"),
     *              @OA\Property(property="updated_at", type="string", example="2023-02-23T12:33:45.000000Z")
     *         )
     *     ),
     *      @OA\Response(
     *          response=404,
     *          description="NOT FOUND",
     *          @OA\JsonContent(
     *              @OA\Property(property="message", type="string", example="No query results for model [App\\Models\\Cliente] #id"),
     *          )
     *      )
     * )
     */
    public function index(Request $request, $id)
    {
        $quotes = Quotes::where('id_entity',$id)->get();
        return response()->json(["data"=>$quotes],200);
    }

     /**
     * @OA\Get (
     *     path="/api/quotes/{id}",
     *     operationId="watch_quotes",
     *     tags={"quotes"},
     *     security={{ "apiAuth": {} }},
     *     summary="See quote",
     *     description="See quote",
     *    @OA\Parameter(
     *         in="path",
     *         name="id",
     *         required=true,
     *         @OA\Schema(type="string")
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="OK",
     *         @OA\JsonContent(
     *              @OA\Property(property="id", type="number", example=1),
     *              @OA\Property(property="name", type="string", example="Process"),
     *              @OA\Property(property="created_at", type="string", example="2023-02-23T00:09:16.000000Z"),
     *              @OA\Property(property="updated_at", type="string", example="2023-02-23T12:33:45.000000Z")
     *         )
     *     ),
     *      @OA\Response(
     *          response=404,
     *          description="NOT FOUND",
     *          @OA\JsonContent(
     *              @OA\Property(property="message", type="string", example="No query results for model [App\\Models\\Cliente] #id"),
     *          )
     *      )
     * )
     */
    public function watch($id){
        try{
            $supplier = Quotes::find($id);
            return response()->json(["data"=>$supplier],200);
        }catch (Exception $e) {
            return response()->json(["data"=>"fail"],200);
        }
    }

    /**
     * @OA\Post(
     *      path="/api/quotes",
     *      operationId="store_quotes",
     *      tags={"quotes"},
     *     security={{ "apiAuth": {} }},
     *      summary="Store quote",
     *      description="Store quote",
     *      @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *            required={"name"},
     *            @OA\Property(property="name", type="string", format="string", example="Name"),
     *            @OA\Property(property="direction", type="string", format="string", example="direction"),
     *            @OA\Property(property="phone", type="string", format="string", example="phone"),
     *         ),
     *      ),
     *     @OA\Response(
     *          response=200, description="Success",
     *          @OA\JsonContent(
     *             @OA\Property(property="status", type="integer", example=""),
     *             @OA\Property(property="data",type="object")
     *          )
     *       )
     *  )
     */
    public function register(Request $request)
    {
        $quotes = new Quotes(request()->all());
        $quotes->save();
        return response()->json(["data"=>$quotes],200);
    }

    /**
     * @OA\Put(
     *      path="/api/quotes/{id}",
     *      operationId="update_quotes",
     *      tags={"quotes"},
     *     security={{ "apiAuth": {} }},
     *      summary="Update quote",
     *      description="Update quote",
     *     @OA\Parameter(
     *         in="path",
     *         name="id",
     *         required=true,
     *         @OA\Schema(type="string")
     *     ),
     *      @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *            required={"name"},
     *            @OA\Property(property="name", type="string", format="string", example="Name"),
     *            @OA\Property(property="direction", type="string", format="string", example="direction"),
     *            @OA\Property(property="phone", type="string", format="string", example="phone"),
     *         ),
     *      ),
     *     @OA\Response(
     *          response=200, description="Success",
     *          @OA\JsonContent(
     *             @OA\Property(property="status", type="integer", example=""),
     *             @OA\Property(property="data",type="object")
     *          )
     *       )
     *  )
     */

    public function update(Request $request, $id){
        try{
            $quotes = Quotes::find($id);
            $quotes->update($request->all());
            return response()->json(["data"=>"ok"],200);
        }catch (Exception $e) {
            return response()->json(["data"=>"fail"],200);
        }
    }

    /**
     * @OA\Delete(
     *      path="/api/quotes/{id}",
     *      operationId="delete_quotes",
     *      tags={"quotes"},
     *     security={{ "apiAuth": {} }},
     *      summary="Delete quote",
     *      description="Delete quote",
     *    @OA\Parameter(
     *         in="path",
     *         name="id",
     *         required=true,
     *         @OA\Schema(type="string")
     *     ),
     *     @OA\Response(
     *          response=200, description="Success",
     *          @OA\JsonContent(
     *             @OA\Property(property="status", type="integer", example=""),
     *             @OA\Property(property="data",type="object")
     *          )
     *       )
     *  )
     */

    public function delete($id){
        try{
            $quotes = Quotes::destroy($id);
            return response()->json(["data"=>"ok"],200);
        }catch (Exception $e) {
            return response()->json(["data"=>"fail"],200);
        }
    }

}